refactor: split GraphQL schema builder files
